# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing bookmarks",
        "title": "Bookmark Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/health-check": {
            "get": {
                "description": "Returns service health status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Check health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/links/redirect/{code}": {
            "get": {
                "description": "Looks up the shortened code in storage and issues an HTTP 302 redirect to the original URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Redirect to original URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shortened code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/links/shorten": {
            "post": {
                "description": "Generates a 7-character alphanumeric code for the given URL and stores it in Redis with the specified TTL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Shorten a URL",
                "parameters": [
                    {
                        "description": "Shorten URL request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ShortenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ShortenRequest": {
            "type": "object",
            "required": [
                "exp",
                "url"
            ],
            "properties": {
                "exp": {
                    "type": "integer",
                    "minimum": 1
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "service.Response": {
            "type": "object",
            "properties": {
                "hostname": {
                    "type": "string"
                },
                "instance_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                }
            }
        }
    }
}